+++ contracts/programs/keystone-forwarder/src/lib.rs
@@ -1,30 +1,22 @@
 use anchor_lang::prelude::*;
 use anchor_lang::solana_program::{hash, keccak, secp256k1_recover::*};
 
-mod error;
+use context::*;
+pub use state::{ForwarderState, ExecutionState, OraclesConfig};
+use utils::{get_config_id, extract_transmission_id};
+use common::{STATE_VERSION, REPORT_CONTEXT_LEN, MAX_ORACLES, SIGNATURE_LEN, FORWARDER_METADATA_LENGTH, METADATA_LENGTH};
 pub use error::*;
 
+mod error;
+mod context;
+mod common;
+mod state;
+mod utils;
+
 declare_id!("whV7Q5pi17hPPyaPksToDw1nMx6Lh8qmNWKFaLRQ4wz");
-
-pub const STATE_VERSION: u8 = 1;
-
-pub const ANCHOR_DISCRIMINATOR: usize = 8;
-
-pub const REPORT_CONTEXT_LEN: usize = 96;
-
-// our don size is directly limited by the transaction size during on_report's signature verification
-pub const MAX_ORACLES: usize = 17;
-
-pub const SIGNATURE_LEN: usize = 65;
-
-pub const FORWARDER_METADATA_LENGTH: usize = 45;
-pub const METADATA_LENGTH: usize = 109;
 #[program]
 pub mod keystone_forwarder {
-    use anchor_lang::{
-        solana_program::{instruction::Instruction, program::invoke_signed, system_instruction},
-        Discriminator,
-    };
+    use anchor_lang::solana_program::{instruction::Instruction, program::invoke_signed};
 
     use super::*;
 
@@ -270,211 +262,6 @@ fn set_oracles_config(
     oracles_config.signer_addresses = signer_addresses;
 
     Ok(())
-}
-
-#[account]
-#[derive(Default, InitSpace)]
-pub struct ForwarderState {
-    pub version: u8,
-    pub authority_nonce: u8, // bump
-    pub owner: Pubkey,
-    pub proposed_owner: Pubkey,
-}
-
-#[derive(Accounts)]
-pub struct Initialize<'info> {
-    // the account is not a PDA but it is initialized by the program
-    #[account(
-        init,
-        payer = owner,
-        space = ANCHOR_DISCRIMINATOR + ForwarderState::INIT_SPACE
-    )]
-    pub state: Account<'info, ForwarderState>,
-    #[account(mut)]
-    pub owner: Signer<'info>,
-
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-pub struct TransferOwnership<'info> {
-    #[account(mut)]
-    pub state: Account<'info, ForwarderState>,
-
-    #[account(address = state.owner @ AuthError::Unauthorized)]
-    pub current_owner: Signer<'info>,
-}
-
-#[derive(Accounts)]
-pub struct AcceptOwnership<'info> {
-    #[account(mut)]
-    pub state: Account<'info, ForwarderState>,
-
-    #[account(address = state.proposed_owner @ AuthError::Unauthorized)]
-    pub proposed_owner: Signer<'info>,
-}
-
-// combination of solidity's OracleSet and the configId mapping
-#[account]
-pub struct OraclesConfig {
-    config_id: u64,
-    f: u8,
-    signer_addresses: Vec<[u8; 20]>,
-}
-
-impl OraclesConfig {
-    pub const INIT_SPACE: usize = 8 + 1 + 4;
-
-    pub fn space_with_signers(num_signers: usize) -> usize {
-        Self::INIT_SPACE + (num_signers * 20)
-    }
-}
-
-fn get_config_id(don_id: u32, config_version: u32) -> u64 {
-    ((don_id as u64) << 32) | (config_version as u64)
-}
-
-#[derive(Accounts)]
-#[instruction(don_id: u32, config_version: u32, f: u8, signer_addresses: Vec<[u8; 20]>)]
-pub struct InitOraclesConfig<'info> {
-    pub state: Account<'info, ForwarderState>,
-
-    #[account(
-        init,
-        payer = owner,
-        seeds = [b"config", state.key().as_ref(), &get_config_id(don_id, config_version).to_be_bytes()],
-        bump,
-        space = ANCHOR_DISCRIMINATOR + OraclesConfig::space_with_signers(signer_addresses.len())
-    )]
-    oracles_config: Account<'info, OraclesConfig>,
-
-    #[account(mut, address = state.owner @ AuthError::Unauthorized)]
-    pub owner: Signer<'info>, // must be the same owner as the one in the state account
-
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-#[instruction(don_id: u32, config_version: u32, f: u8, signer_addresses: Vec<[u8; 20]>)]
-pub struct UpdateOraclesConfig<'info> {
-    pub state: Account<'info, ForwarderState>,
-
-    #[account(
-        mut,
-        seeds = [b"config", state.key().as_ref(), &get_config_id(don_id, config_version).to_be_bytes()],
-        bump,
-        realloc = ANCHOR_DISCRIMINATOR + OraclesConfig::space_with_signers(signer_addresses.len()),
-        realloc::payer = owner,
-        realloc::zero = true
-    )]
-    oracles_config: Account<'info, OraclesConfig>,
-
-    #[account(mut, address = state.owner @ AuthError::Unauthorized)]
-    pub owner: Signer<'info>,
-
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-#[instruction(don_id: u32, config_version: u32)]
-pub struct CloseOraclesConfig<'info> {
-    pub state: Account<'info, ForwarderState>,
-
-    #[account(
-        mut,
-        seeds = [b"config", state.key().as_ref(), &get_config_id(don_id, config_version).to_be_bytes()],
-        bump,
-        close = owner
-    )]
-    oracles_config: Account<'info, OraclesConfig>,
-
-    #[account(mut, address = state.owner @ AuthError::Unauthorized)]
-    pub owner: Signer<'info>, // must be the same owner as the one in the state account
-}
-
-// data =  len_signatures (1) | signatures (N*65) | raw_report (M) | report_context (96)
-fn extract_raw_report(data: &[u8]) -> &[u8] {
-    let num_signatures = data[0] as usize;
-    let data = &data[1..];
-    let _signatures = &data[..num_signatures * SIGNATURE_LEN];
-    let data = &data[num_signatures * SIGNATURE_LEN..];
-    let _report_context = &data[data.len() - REPORT_CONTEXT_LEN..];
-
-    // raw report
-    &data[..data.len() - REPORT_CONTEXT_LEN]
-}
-
-// version                offset   0, size  1
-// workflow_execution_id  offset   1, size 32
-// timestamp              offset  33, size  4
-// don_id                 offset  37, size  4
-// don_config_version     offset  41, size  4
-// workflow_cid           offset  45, size 32
-// workflow_name          offset  77, size 10
-// workflow_owner         offset  87, size 20
-// report_id              offset 107, size  2
-
-fn extract_config_id(raw_report: &[u8]) -> [u8; 8] {
-    // don_id | don_config_version
-    raw_report[37..45].try_into().expect("Expected 8 bytes")
-}
-
-fn extract_transmission_id(raw_report: &[u8], receiver: &Pubkey) -> [u8; 32] {
-    let workflow_execution_id = &raw_report[1..33];
-    let report_id = &raw_report[107..109];
-
-    // use sha-256 bc it's much cheaper than keccak-256
-    hash::hash(&[&receiver.to_bytes(), workflow_execution_id, report_id].concat()).to_bytes()
-}
-
-#[derive(Accounts)]
-#[instruction(data: Vec<u8>)]
-pub struct Report<'info> {
-    pub state: Account<'info, ForwarderState>,
-
-    #[account(
-        mut,
-        seeds = [b"config", state.key().as_ref(), &extract_config_id(extract_raw_report(&data))],
-        bump
-    )]
-    oracles_config: Account<'info, OraclesConfig>,
-
-    #[account(mut)]
-    pub transmitter: Signer<'info>,
-
-    /// CHECK: This is a PDA
-    #[account(seeds = [b"forwarder", state.key().as_ref()], bump = state.authority_nonce)]
-    pub forwarder_authority: UncheckedAccount<'info>,
-
-    // it is dependent on the state.key(), a predetermined bump, workflow execution id, config_id, report_id
-    #[account(
-        init_if_needed,
-        payer = transmitter,
-        space = ANCHOR_DISCRIMINATOR + ExecutionState::INIT_SPACE,
-        seeds = [
-            b"execution_state", 
-            state.key().as_ref(),
-            &extract_transmission_id(extract_raw_report(&data), receiver_program.key)
-        ],
-        bump
-    )]
-    pub execution_state: Account<'info, ExecutionState>,
-
-    #[account(executable)]
-    /// CHECK: We don't use Program<> here since it can be any program, "executable" is enough
-    pub receiver_program: UncheckedAccount<'info>,
-
-    pub system_program: Program<'info, System>,
-    // remaining accounts passed to receiver
-}
-
-#[account]
-#[derive(Default, InitSpace)]
-pub struct ExecutionState {
-    pub transmitter: Pubkey,
-    pub transmission_id: [u8; 32],
-    // until failure states are reported by the write target, success will always be true
-    pub success: bool,
 }
 
 //
